requirements.txt
Pillow>=9.0.0
openpyxl>=3.0.0
pillow-heif>=0.10.0
pyinstaller>=5.0.0